requirements.txt
streamlit
torch
diffusers
transformers
opencv-python
numpy
Pillow
bpy
